[생성] 탭바에 타임 라인 연결 [수정] 날씨데이터 저장 수정 	- weatherId 지정 [수정] 타임라인 셀에 엔트리 데이터 연결 [수정] 타임라인 상단 뷰 조정 	- navigationBar 경계삭제 	- 아래로 당길시 나타나는 테이블 뷰 상단 배경 채우기
