+++ OneDay/OneDay/Views/Storyboards/Base.lproj/Main.storyboard
@@ -44,20 +44,14 @@
             <point key="canvasLocation" x="1052" y="-287"/>
         </scene>
         <!--Timeline-->
-        <scene sceneID="tne-QT-ifu">
+        <scene sceneID="aSf-8X-AG0">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="TimeLineViewController" customModule="OneDay" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="551"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
-                    </view>
-                    <navigationItem key="navigationItem" title="Timeline" id="DNz-Pb-dhb"/>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
+                <viewControllerPlaceholder storyboardName="Coredata" referencedIdentifier="Timeline" id="8yw-PF-Uz6" sceneMemberID="viewController">
+                    <navigationItem key="navigationItem" id="Goh-VP-ypY"/>
+                </viewControllerPlaceholder>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="K7v-Ps-o3s" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="1991" y="-1681"/>
+            <point key="canvasLocation" x="1931" y="-1681"/>
         </scene>
         <!--Map-->
         <scene sceneID="chG-y2-ejf">
@@ -125,13 +119,16 @@
                         <rect key="frame" x="0.0" y="20" width="375" height="96"/>
                         <autoresizingMask key="autoresizingMask"/>
                         <color key="barTintColor" name="doBlue"/>
+                        <textAttributes key="titleTextAttributes">
+                            <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        </textAttributes>
                         <textAttributes key="largeTitleTextAttributes">
                             <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                         </textAttributes>
                     </navigationBar>
                     <nil name="viewControllers"/>
                     <connections>
-                        <segue destination="BYZ-38-t0r" kind="relationship" relationship="rootViewController" id="8vK-wQ-Qsp"/>
+                        <segue destination="8yw-PF-Uz6" kind="relationship" relationship="rootViewController" id="u99-AA-QNy"/>
                     </connections>
                 </navigationController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="8st-5g-w6C" userLabel="First Responder" sceneMemberID="firstResponder"/>
@@ -215,11 +212,11 @@
         </scene>
     </scenes>
     <resources>
-        <image name="add" width="36.799999237060547" height="36.799999237060547"/>
-        <image name="ic_calendar" width="34.5" height="34.5"/>
-        <image name="map" width="36.799999237060547" height="36.799999237060547"/>
-        <image name="picture" width="27" height="27"/>
-        <image name="tabs" width="36.799999237060547" height="36.799999237060547"/>
+        <image name="add" width="23" height="23"/>
+        <image name="ic_calendar" width="23" height="23"/>
+        <image name="map" width="23" height="23"/>
+        <image name="picture" width="36" height="36"/>
+        <image name="tabs" width="23" height="23"/>
         <namedColor name="doBlue">
             <color red="0.41176470588235292" green="0.74509803921568629" blue="0.98039215686274506" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </namedColor>
